requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.17.0
oauthlib==3.0.2
PySocks==1.7.0
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
scikit-learn==0.21.3
scipy==1.3.0
six==1.12.0
tweepy==3.8.0
urllib3==1.25.3
Werkzeug==0.15.5
